+++ banchan/src/components/StateProvider.jsx
@@ -1,9 +1,8 @@
 import { useState } from "react";
 import Header from "./header/Header";
 import MainPage from "./main/MainPage";
-import styled, { injectGlobal } from "styled-components";
 
-const StateProvider = (props) => {
+const StateProvider = () => {
   const [loginState, setLoginState] = useState(false);
   // login해서 상태값 받아와야 함
