+++ banchan/src/components/main/CarouselSection.jsx
@@ -1,17 +1,17 @@
-import { useEffect, useState } from 'react';
-import Card from '../utils/Card';
-import theme from '../utils/constant';
+import { useEffect, useState } from "react";
+import Card from "../utils/Card";
+import theme from "../utils/theme";
 
 const CarouselSection = (props) => {
   const [products, setProducts] = useState([]);
 
   useEffect(() => {
     fetch(
-      'https://h3rb9c0ugl.execute-api.ap-northeast-2.amazonaws.com/develop/baminchan/main'
+      "https://h3rb9c0ugl.execute-api.ap-northeast-2.amazonaws.com/develop/baminchan/main"
     )
       .then((response) => response.json())
       .then((result) => setProducts(result.body))
-      .then((error) => console.log('error', error));
+      .then((error) => console.log("error", error));
   }, []);
 
   return (
