+++ banchan/src/components/utils/Price.jsx
@@ -1,5 +1,5 @@
-import styled from 'styled-components';
-import theme from './constant';
+import styled from "styled-components";
+import theme from "./theme";
 
 const Price = ({ product }) => {
   const isOnSale = !!product.n_price;
