+++ banchan/src/components/utils/theme.js
@@ -1,0 +1,28 @@
+const theme = {
+  colors: {
+    darkGray: "#333",
+    gray: "#828282",
+    lightGray: "#bdbdbd",
+    lightGrayBG: "#F5F5F7",
+    whiteBlue: "#EEF4FA",
+    green: "#82D32D",
+    lightBlue: "#86C6FF",
+    white: "#FFF",
+  },
+  fontSizes: {
+    XS: "14px",
+    S: "16px",
+    M: "18px",
+    L: "20px",
+    XL: "24px",
+  },
+  cardSizes: {
+    M: "308px",
+    L: "384px",
+  },
+  borders: {
+    radius: "5px",
+  },
+};
+
+export default theme;
